Add files via upload
An array is filled with words from the dictionary.txt .
This array is sent to a method that fills a HashTable using double hashing with the words in array (each word = a specific int value based on their characters and index).
Another hashtable is created that fills it with words in a randomized manner. 
Indexes of words are checked from mytable (Hashtable) to the randomized table, if not found it's a collision.